(feat) process args with serial refconfig
