+++ config/args/args_test.go
@@ -6,7 +6,7 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
-	"github.com/jahvon/flow/cmd/internal/args"
+	"github.com/jahvon/flow/config/args"
 )
 
 func TestParseArgs(t *testing.T) {
